refactor: add a hook for api consumption
