+++ electronic-booklet/src/interface/medicine.ts
@@ -2,3 +2,7 @@ export interface Medicine {
     id: string
     name: string
 }
+
+export interface MedicineProps {
+    medicine: Medicine[]
+}
