+++ electronic-booklet/src/interface/page-button-props.ts
@@ -1,0 +1,5 @@
+export interface PageButtonProps {
+    pageNumber: number
+    isActive: boolean
+    onClick: () => void
+}
